✨ Feat :  JinniContext에 color 추가
