+++ src/constants/breakpoint.ts
@@ -1,0 +1,1 @@
+export const BREAKPOINTS = ['xs', 'sm', 'md', 'lg', 'xl'] as const;
