--- src/constants/css-color-properties.ts
@@ -1,11 +1,0 @@
-export const CSS_COLOR_PROPERTIES = [
-  'color',
-  'backgroundColor',
-  'borderColor',
-  'outlineColor',
-  'textDecorationColor',
-  'caretColor',
-  'columnRuleColor',
-  'fill',
-  'stroke'
-] as const;
