+++ src/contexts/JinniContext.ts
@@ -1,8 +1,13 @@
 import { createContext } from 'react';
 import { BreakpointType } from '@/types/breakpoint';
+import { JinniColorTheme, JinniColorPalette } from '@/types/color';
 
-interface JinniValueType {
+export interface JinniValueType {
   breakpoints: Record<BreakpointType, number>;
+  color: {
+    theme: Record<JinniColorTheme, string>;
+    palette: Record<JinniColorPalette, string>;
+  };
 }
 
 const JinniContext = createContext<JinniValueType | null>(null);
