+++ src/types/breakpoint.ts
@@ -1,3 +1,5 @@
-export type BreakpointType = 'xs' | 'sm' | 'md' | 'lg' | 'xl';
+import { BREAKPOINTS } from '@/constants/breakpoint';
+
+export type BreakpointType = (typeof BREAKPOINTS)[number];
 
 export type Responsive<T> = Partial<Record<BreakpointType, T>>;
