+++ src/types/style.ts
@@ -1,5 +1,5 @@
 import { ColorType } from '@/types/color';
-import { CSS_COLOR_PROPERTIES } from '@/constants/css-color-properties';
+import { CSS_COLOR_PROPERTIES } from '@/constants/color';
 import { Responsive } from '@/types/breakpoint';
 
 type CSSColorProperties = (typeof CSS_COLOR_PROPERTIES)[number];
