+++ src/utils/isResponsive.ts
@@ -1,8 +1,9 @@
 import { Responsive } from '@/types/breakpoint';
+import { BREAKPOINTS } from '@/constants/breakpoint';
 
 export const isResponsive = <T>(element: unknown): element is Responsive<T> => {
   if (!element || typeof element !== 'object') return false;
-
-  const breakpoints = ['xs', 'sm', 'md', 'lg', 'xl'];
-  return Object.keys(element).every((key) => breakpoints.includes(key));
+  return Object.keys(element).every((key) =>
+    BREAKPOINTS.some((bp) => bp === key)
+  );
 };
